move check response to utils
