+++ cohere/utils.py
@@ -3,9 +3,10 @@ import asyncio
 import json
 import time
 import typing
-from typing import Awaitable, Callable, Optional, TypeVar
+from typing import Awaitable, Callable, Dict, Optional, TypeVar
 
-from cohere.error import CohereError
+from cohere.error import CohereAPIError, CohereError
+from cohere.logging import logger
 
 try:  # numpy is optional, but support json encoding if the user has it
     import numpy as np
@@ -94,3 +95,22 @@ async def async_wait_for_job(
         job = await get_job()
 
     return job
+
+
+def check_response(json_response: Dict, headers: Dict, status_code: int):
+    if "X-API-Warning" in headers:
+        logger.warning(headers["X-API-Warning"])
+    if "message" in json_response:  # has errors
+        raise CohereAPIError(
+            message=json_response["message"],
+            http_status=status_code,
+            headers=headers,
+        )
+    if 400 <= status_code < 500:
+        raise CohereAPIError(
+            message=f"Unexpected client error (status {status_code}): {json_response}",
+            http_status=status_code,
+            headers=headers,
+        )
+    if status_code >= 500:
+        raise CohereError(message=f"Unexpected server error (status {status_code}): {json_response}")
